+++ 8.1-litespeed/root/60-php-config.sh
@@ -1,4 +1,7 @@
-#!/bin/bash
+#!/usr/bin/env bash
+
+set -e
+
 
 echo >&2 "Setting PHP tunables..."