+++ 8.2-litespeed/root/40-ls-config.sh
@@ -1,4 +1,7 @@
-#!/bin/bash
+#!/usr/bin/env bash
+
+set -e
+
 
 if ! [ "$(ls -A /usr/local/lsws)" ]; then
   echo >&2 "OpenLiteSpeed not configured, setting configuration..."